rename test_properties to test_fields_properties
